# btf: Optimize string table for globally increasing offsets