@lmb 2023-11-07
I get this on my machine:

```
$ benchstat base.txt opt.txt 
goos: linux
goarch: amd64
pkg: github.com/cilium/ebpf/btf
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                │  base.txt   │              opt.txt              │
                │   sec/op    │   sec/op     vs base              │
ParseVmlinux-16   42.76m ± 3%   40.35m ± 2%  -5.66% (p=0.002 n=6)

                │   base.txt   │            opt.txt            │
                │     B/op     │     B/op      vs base         │
ParseVmlinux-16   31.45Mi ± 0%   31.45Mi ± 0%  ~ (p=0.132 n=6)

                │  base.txt   │           opt.txt            │
                │  allocs/op  │  allocs/op   vs base         │
ParseVmlinux-16   534.1k ± 0%   534.1k ± 0%  ~ (p=0.165 n=6)
```

The laptop is plugged in and I ran `sudo cpupower frequency-set -g performance`.